+++ runtime/fireworks/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 edition = "2018"
-name = "fireworks-runtime"
+name = "snd-fireworks-ctl-service"
 version = "0.1.0"
 # For publishing.
 authors = ["Takashi Sakamoto"]
 description = """
-Implementation of service runtime for Fireworks board module
+Runtime implementation to control audio and music unit in IEEE 1394 bus which ALSA fireworks driver
+(snd-fireworks) supports.
 """
 license = "GPL-3.0-or-later"
 publish = false
